Moved udp tests to its own file.
