+++ test/common/network/BUILD
@@ -161,6 +161,24 @@ envoy_cc_test(
 )
 
 envoy_cc_test(
+    name = "udp_listener_impl_test",
+    srcs = ["udp_listener_impl_test.cc"],
+    deps = [
+        "//source/common/event:dispatcher_lib",
+        "//source/common/network:address_lib",
+        "//source/common/network:listener_lib",
+        "//source/common/network:utility_lib",
+        "//source/common/stats:stats_lib",
+        "//test/mocks/network:network_mocks",
+        "//test/mocks/server:server_mocks",
+        "//test/test_common:environment_lib",
+        "//test/test_common:network_utility_lib",
+        "//test/test_common:test_time_lib",
+        "//test/test_common:utility_lib",
+    ],
+)
+
+envoy_cc_test(
     name = "resolver_test",
     srcs = ["resolver_impl_test.cc"],
     deps = [
